Fix issue with modifying text inside existing spellcheck spans
When text inside spellcheck spans was modified, the span was unwrapped
so that the error styling would be removed.  However, this detached the
span from the DOM and therefore had no text.  This empty element then
triggered a whole-page spellcheck as it was passed through the
spellcheck events.  I added a function to find the nearest parent block
element (which will probably be helpful to have later as well), and set
the passed element to that value instead